third_party/blink/web_tests/external/wpt/dom/idlharness.window_exclude=Node-expected.txt

This is a testharness.js-based test.
Found 22 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Document interface: attribute customElementRegistry
  assert_true: The prototype object must have a property "customElementRegistry" expected true got false
[FAIL] Document interface: new Document() must inherit property "customElementRegistry" with the proper type
  assert_inherits: property "customElementRegistry" not found in prototype chain
[FAIL] Document interface: xmlDoc must inherit property "customElementRegistry" with the proper type
  assert_inherits: property "customElementRegistry" not found in prototype chain
[FAIL] ShadowRoot interface: attribute customElementRegistry
  assert_true: The prototype object must have a property "customElementRegistry" expected true got false
[FAIL] Element interface: attribute customElementRegistry
  assert_true: The prototype object must have a property "customElementRegistry" expected true got false
[FAIL] Element interface: element must inherit property "customElementRegistry" with the proper type
  assert_inherits: property "customElementRegistry" not found in prototype chain
[FAIL] XSLTProcessor must be primary interface of new XSLTProcessor()
  assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled"
[FAIL] Stringification of new XSLTProcessor()
  assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled"
[FAIL] XSLTProcessor interface: new XSLTProcessor() must inherit property "importStylesheet(Node)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled"
[FAIL] XSLTProcessor interface: calling importStylesheet(Node) on new XSLTProcessor() with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled"
[FAIL] XSLTProcessor interface: new XSLTProcessor() must inherit property "transformToFragment(Node, Document)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled"
[FAIL] XSLTProcessor interface: calling transformToFragment(Node, Document) on new XSLTProcessor() with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled"
[FAIL] XSLTProcessor interface: new XSLTProcessor() must inherit property "transformToDocument(Node)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled"
[FAIL] XSLTProcessor interface: calling transformToDocument(Node) on new XSLTProcessor() with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled"
[FAIL] XSLTProcessor interface: new XSLTProcessor() must inherit property "setParameter(DOMString, DOMString, any)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled"
[FAIL] XSLTProcessor interface: calling setParameter(DOMString, DOMString, any) on new XSLTProcessor() with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled"
[FAIL] XSLTProcessor interface: new XSLTProcessor() must inherit property "getParameter(DOMString, DOMString)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled"
[FAIL] XSLTProcessor interface: calling getParameter(DOMString, DOMString) on new XSLTProcessor() with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled"
[FAIL] XSLTProcessor interface: new XSLTProcessor() must inherit property "removeParameter(DOMString, DOMString)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled"
[FAIL] XSLTProcessor interface: calling removeParameter(DOMString, DOMString) on new XSLTProcessor() with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled"
[FAIL] XSLTProcessor interface: new XSLTProcessor() must inherit property "clearParameters()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled"
[FAIL] XSLTProcessor interface: new XSLTProcessor() must inherit property "reset()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled"
Harness: the test ran to completion.
